added iterate functionality and syntax macro
This is something like a mixture of pythons iterators, streams and
racket's generators, but in a more functional way. The main idea
came from racket's generators and implementation unexpectedly showed
similarity to streams. The for syntax is something like an extended
let, which automagically steps through the iterators by itself.

Note: Documentation is not complete yet.